+++ packages/ladle/lib/app/src/addons/theme.tsx
@@ -1,4 +1,6 @@
 import queryString from "query-string";
+import { useHotkeys } from "react-hotkeys-hook";
+import config from "../get-config";
 import { Bulb } from "../icons";
 import { ThemeState, AddonProps, ActionType } from "../../../shared/types";
 
@@ -19,6 +21,17 @@ export const getQuery = (locationSearch: string) => {
 export const Button = ({ globalState, dispatch }: AddonProps) => {
   const darkText = "Switch to dark theme.";
   const lightText = "Switch to light theme.";
+  const changeTheme = () => {
+    const newTheme =
+      globalState.theme === ThemeState.Light
+        ? ThemeState.Dark
+        : ThemeState.Light;
+    document.documentElement.setAttribute("data-theme", newTheme);
+    dispatch({ type: ActionType.UpdateTheme, value: newTheme });
+  };
+  useHotkeys(config.hotkeys.darkMode, changeTheme, {
+    enabled: globalState.hotkeys && config.addons.mode.enabled,
+  });
   return (
     <li>
       <button
@@ -26,14 +39,7 @@ export const Button = ({ globalState, dispatch }: AddonProps) => {
           globalState.theme === ThemeState.Light ? darkText : lightText
         }
         title={globalState.theme === ThemeState.Light ? darkText : lightText}
-        onClick={() => {
-          const newTheme =
-            globalState.theme === ThemeState.Light
-              ? ThemeState.Dark
-              : ThemeState.Light;
-          document.documentElement.setAttribute("data-theme", newTheme);
-          dispatch({ type: ActionType.UpdateTheme, value: newTheme });
-        }}
+        onClick={changeTheme}
         type="button"
       >
         <Bulb />